# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BAC MASTER CI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # USERS
...
  /exercises:
    get:
      operationId: getExercises
      tags: [exercises]
      summary: Get exercises
      parameters:
        - name: lessonId
          in: query
          schema:
            type: integer
        - name: subjectId
          in: query
          schema:
            type: integer
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [easy, medium, hard]
        - name: series
          in: query
          schema:
            type: string
            enum: [A, C, D]
      responses:
        "200":
          description: List of exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"
    post:
      operationId: createExercise
      tags: [exercises]
      summary: Create an exercise (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExerciseRequest"
      responses:
        "201":
          description: Created exercise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
  /exercises/{id}:
    get:
      operationId: getExercise
      tags: [exercises]
      summary: Get exercise by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exercise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
  /exercises/{id}/submit:
    post:
      operationId: submitExercise
      tags: [exercises]
      summary: Submit exercise answer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitExerciseRequest"
      responses:
        "200":
          description: Exercise result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseResult"

  # PAST EXAMS
  /exams:
    get:
      operationId: getExams
      tags: [exams]
      summary: Get past exams
      parameters:
        - name: series
          in: query
          schema:
            type: string
            enum: [A, C, D]
        - name: year
          in: query
          schema:
            type: integer
        - name: subjectId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of exams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exam"
    post:
      operationId: createExam
      tags: [exams]
      summary: Create a past exam (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExamRequest"
      responses:
        "201":
          description: Created exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
  /exams/{id}:
    get:
      operationId: getExam
      tags: [exams]
      summary: Get exam by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"

  # PROGRESS
...
  /chat/message:
    post:
      operationId: sendChatMessage
      tags: [chatbot]
      summary: Send a message to the AI tutor
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageRequest"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessageResponse"
  /chat/history:
    get:
      operationId: getChatHistory
      tags: [chatbot]
      summary: Get chat history
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"

  # ADMIN
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        series:
          type: string
          enum: [A, C, D]
      required:
        - name
        - email
        - password
        - series
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        series:
          type: string
          enum: [A, C, D]
        role:
          type: string
          enum: [student, admin]
        avatarUrl:
          type: string
          nullable: true
        points:
          type: integer
        isPremium:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - series
        - role
        - points
        - isPremium
        - createdAt
...
    Exercise:
      type: object
      properties:
        id:
          type: integer
        lessonId:
          type: integer
          nullable: true
        subjectId:
          type: integer
        series:
          type: string
          enum: [A, C, D, ALL]
        question:
          type: string
        type:
          type: string
          enum: [mcq, open, true_false]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        options:
          type: array
          nullable: true
          items:
            type: string
        correctAnswer:
          type: string
        explanation:
          type: string
        isPremium:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - subjectId
        - series
        - question
        - type
        - difficulty
        - correctAnswer
        - explanation
        - isPremium
        - createdAt
    CreateExerciseRequest:
      type: object
      properties:
        lessonId:
          type: integer
        subjectId:
          type: integer
        series:
          type: string
          enum: [A, C, D, ALL]
        question:
          type: string
        type:
          type: string
          enum: [mcq, open, true_false]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        explanation:
          type: string
        isPremium:
          type: boolean
      required:
        - subjectId
        - series
        - question
        - type
        - difficulty
        - correctAnswer
        - explanation
    SubmitExerciseRequest:
      type: object
      properties:
        answer:
          type: string
      required:
        - answer
    ExerciseResult:
      type: object
      properties:
        correct:
          type: boolean
        explanation:
          type: string
        correctAnswer:
          type: string
        pointsEarned:
          type: integer
      required:
        - correct
        - explanation
        - correctAnswer
        - pointsEarned
    Exam:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        subjectId:
          type: integer
        subjectName:
          type: string
        series:
          type: string
          enum: [A, C, D]
        year:
          type: integer
        content:
          type: string
        correction:
          type: string
        pdfUrl:
          type: string
          nullable: true
        isPremium:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - subjectId
        - subjectName
        - series
        - year
        - content
        - correction
        - isPremium
        - createdAt
    CreateExamRequest:
      type: object
      properties:
        title:
          type: string
        subjectId:
          type: integer
        series:
          type: string
          enum: [A, C, D]
        year:
          type: integer
        content:
          type: string
        correction:
          type: string
        pdfUrl:
          type: string
        isPremium:
          type: boolean
      required:
        - title
        - subjectId
        - series
        - year
        - content
        - correction
...
    ChatMessageRequest:
      type: object
      properties:
        message:
          type: string
        context:
          type: string
          nullable: true
          description: Optional context (lesson content, subject)
      required:
        - message
    ChatMessageResponse:
      type: object
      properties:
        reply:
          type: string
        isEducational:
          type: boolean
      required:
        - reply
        - isEducational
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - role
        - content
        - createdAt
...