# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BAC MASTER CI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";
import type { UserSeries } from "./userSeries";

export interface User {
  id: number;
  name: string;
  email: string;
  series: UserSeries;
  role: UserRole;
  avatarUrl?: string | null;
  points: number;
  isPremium: boolean;
  createdAt: Date;
}